Added Edit Mark Activity, Fixed Bugs
Fixed issues related to scroll view not scrolling both directions at the
same time. Added in a new View that allows list views to have a maximum
height attribute to them. Added in the ability to load the marks that
have been placed thus far, so that they can be edited or deleted by the
user.