Edits now work.
Bit of a bad news; Glide doesn't work well when it comes to it's
callbacks on when it is done. On the other hand, Picasso does a great
job here; so I brought in Picasso. After using it, I realize that Picasso can't load bitmaps, so I need both libraries for the time being until I can figure out an adequate solution.